chore(fe/module/flashcards): Update default flashcard card display

+++ shared/rust/src/domain/jig/module/body/flashcards.rs
@@ -44,7 +44,7 @@ pub enum DisplayMode {
 
 impl Default for DisplayMode {
     fn default() -> Self {
-        Self::Single
+        Self::Double
     }
 }
